function for seat data
