+++ src/components/Breadcrumb/stylesheet.scss
@@ -3,7 +3,7 @@
 .Breadcrumb {
   display: flex;
   align-items: center;
-  font-size: 16px;
+  font-size: 1em;
   padding: 8px 12px;
 
   .list {
@@ -28,7 +28,7 @@
 
       .separator {
         margin: 0 6px;
-        font-size: 12px;
+        font-size: 0.75em;
       }
     }
   }
