+++ src/components/SeatInfo/index.tsx
@@ -3,59 +3,14 @@ import useSWR from 'swr';
 
 import { getContentClassName, getLabelClassName } from '../../utils/misc';
 import { Section as SectionBean } from '../../data/beans';
+import { OccupiedInfo, SeatData } from '../../data/beans/Section';
 
 import './stylesheet.scss';
-
-// Seating information
-export type OccupiedInfo = {
-  occupied: number;
-  total: number;
-};
 
 export type SeatInfoProps = {
   section: SectionBean;
   term: string;
   color: string | undefined;
-};
-
-// Response from fetchSeating
-type SeatData = {
-  inClass: OccupiedInfo | null;
-  waitlist: OccupiedInfo | null;
-};
-
-// Fetch and normalize seating data for a section/term
-const fetchSeating = async (
-  section: SectionBean,
-  term: string
-): Promise<SeatData> => {
-  try {
-    const raw = await section.fetchSeating(term);
-
-    // Handle missing or bad data, assuming less than 4 return values is invalid
-    if (!raw[0] || raw[0].length < 4) {
-      return { inClass: null, waitlist: null };
-    }
-
-    const [inClassTotal, inClassOccupied, waitlistTotal, waitlistOccupied] =
-      raw[0];
-
-    // normalize raw values into an OccupiedInfo object
-    const toOccupiedInfo = (
-      total: unknown,
-      occupied: unknown
-    ): OccupiedInfo => ({
-      occupied: Number(occupied ?? 0),
-      total: Number(total ?? 0),
-    });
-
-    return {
-      inClass: toOccupiedInfo(inClassTotal, inClassOccupied),
-      waitlist: toOccupiedInfo(waitlistTotal, waitlistOccupied),
-    };
-  } catch (err) {
-    return { inClass: null, waitlist: null };
-  }
 };
 
 // subcomponent for a single stat line (label + numbers)
@@ -93,7 +48,7 @@ export default function SeatInfo({
 }: SeatInfoProps): React.ReactElement {
   const { data: seatData, isLoading } = useSWR<SeatData>(
     ['seating', section.crn, term],
-    () => fetchSeating(section, term)
+    () => section.getSeatData(term)
   );
 
   // Derive styling classes based on Section's color (passed down from Section)
